update #3
Добавлено:
- Работает стрельба в разные стороны (По направлению курсора мыши)
- Дроп предметов, добавление в инвентарь, использование в из инвентаря.
- Основной инвентарь (Кнопка E)
- Вероятность попасть по противнику (weapon.hit_chance)
- окно диалога

- ДОБАВИЛ Я ЗВУК СТРЕЛЬБЫ, ПАВЕЛ ВАЛЕРЬЕВИЧ)))))))

+++ core/src/com/samsung/game/engine/Collideable.java
@@ -1,8 +1,8 @@
 package com.samsung.game.engine;
 
 public interface Collideable {
-    public int getWidth();
-    public int getHeight();
+    public float getWidth();
+    public float getHeight();
     public float getX();
     public float getY();
